♻️ refactor(server): extract BookFtsWriter and helpers from BookRepository
